requirements.txt
CairoSVG==2.7.1
clip==1.0
faiss_cpu==1.8.0.post1
lxml==5.2.2
numpy==2.1.2
openai==1.51.2
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
torch==2.4.0
tqdm==4.65.0
transformers==4.43.3
